pyproject.toml
[project]
name = "agent-framework-graphrag-demo"
version = "0.1.0"
description = "GraphRAG contract review demo using Microsoft Agent Framework and Neo4j"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework>=0.1.0",
    "azure-identity>=1.19.0",
    "neo4j>=5.27.0",
    "neo4j-graphrag>=0.1.0",
    "openai>=1.58.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
